# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Coherence API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/v1/axes/build": {
      "post": {
        "tags": [
          "axes"
        ],
        "summary": "Build",
        "operationId": "build_v1_axes_build_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ethicalai__api__axes__BuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/axes/{pack_id}/activate": {
      "post": {
        "tags": [
          "axes"
        ],
        "summary": "Activate Pack",
        "operationId": "activate_pack_v1_axes__pack_id__activate_post",
        "parameters": [
          {
            "name": "pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/axes/{pack_id}": {
      "get": {
        "tags": [
          "axes"
        ],
        "summary": "Get Pack",
        "operationId": "get_pack_v1_axes__pack_id__get",
        "parameters": [
          {
            "name": "pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/axes/{pack_id}/export": {
      "get": {
        "tags": [
          "axes"
        ],
        "summary": "Export Pack",
        "description": "Export full axis pack vectors as JSON for inline testing.\n\nNote: Large payloads; intended for dev/test only.",
        "operationId": "export_pack_v1_axes__pack_id__export_get",
        "parameters": [
          {
            "name": "pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/axes/{axis_pack_id}": {
      "get": {
        "tags": [
          "axes"
        ],
        "summary": "Get Pack",
        "description": "Get a single axis pack summary by ID.",
        "operationId": "get_pack_axes__axis_pack_id__get",
        "parameters": [
          {
            "name": "axis_pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Axis Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Pack Axes  Axis Pack Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/axes/create": {
      "post": {
        "tags": [
          "axes"
        ],
        "summary": "Create Pack",
        "description": "Create an axis pack from seed phrases and persist it.\n\nUses the default encoder and diff-of-means builder.\nSaves to data/axes/{axis_pack_id}.json",
        "operationId": "create_pack_axes_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAxisPack"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAxisPackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivateResponse": {
        "properties": {
          "active": {
            "additionalProperties": true,
            "type": "object",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "active"
        ],
        "title": "ActivateResponse"
      },
...
      "CreateAxisPack": {
        "properties": {
          "axes": {
            "items": {
              "$ref": "#/components/schemas/AxisSeed"
            },
            "type": "array",
            "title": "Axes"
          },
          "method": {
            "type": "string",
            "enum": [
              "diffmean",
              "cca",
              "lda"
            ],
            "title": "Method",
            "default": "diffmean"
          },
          "choquet_capacity": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Choquet Capacity"
          },
          "lambda_": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lambda"
          },
          "beta": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beta"
          },
          "weights": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights"
          }
        },
        "type": "object",
        "required": [
          "axes"
        ],
        "title": "CreateAxisPack",
        "description": "Create an axis pack request.\n\n- axes: list of `AxisSeed`\n- method: \"diffmean\" | \"cca\" | \"lda\"\n- choquet_capacity: optional capacity mapping (keys like \"0,2,3\")\n- lambda_/beta/weights: optional per-axis parameters"
      },
      "CreateAxisPackResponse": {
        "properties": {
          "axis_pack_id": {
            "type": "string",
            "title": "Axis Pack Id"
          },
          "k": {
            "type": "integer",
            "title": "K"
          },
          "names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Names"
          }
        },
        "type": "object",
        "required": [
          "axis_pack_id",
          "k",
          "names"
        ],
        "title": "CreateAxisPackResponse",
        "description": "Response after creating an axis pack."
      },
...
      "ExportResponse": {
        "properties": {
          "pack_id": {
            "type": "string",
            "title": "Pack Id"
          },
          "names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Names"
          },
          "Q": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Q"
          },
          "lambda_": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Lambda"
          },
          "beta": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Beta"
          },
          "weights": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Weights"
          }
        },
        "type": "object",
        "required": [
          "pack_id",
          "names",
          "Q",
          "lambda_",
          "beta",
          "weights"
        ],
        "title": "ExportResponse"
      },
...
      "GetResponse": {
        "properties": {
          "pack_id": {
            "type": "string",
            "title": "Pack Id"
          },
          "dim": {
            "type": "integer",
            "title": "Dim"
          },
          "k": {
            "type": "integer",
            "title": "K"
          },
          "names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Names"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          },
          "pack_hash": {
            "type": "string",
            "title": "Pack Hash"
          }
        },
        "type": "object",
        "required": [
          "pack_id",
          "dim",
          "k",
          "names",
          "meta",
          "pack_hash"
        ],
        "title": "GetResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ethicalai__api__axes__BuildRequest": {
        "properties": {
          "names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Names"
          },
          "seed_phrases": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed Phrases"
          },
          "seed_vectors": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed Vectors"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta",
            "default": {}
          },
          "thresholds_data": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "items": {
                      "type": "number"
                    },
                    "type": "array"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thresholds Data"
          },
          "fpr_max": {
            "type": "number",
            "title": "Fpr Max",
            "default": 0.05
          }
        },
        "type": "object",
        "required": [
          "names"
        ],
        "title": "BuildRequest"
      }
    }
  }
}